backend/model_integration.py
# This is a placeholder file to show how you might integrate your AI model
import numpy as np
from PIL import Image
import io
import base64
from typing import Dict, Any, Optional

# This would be your actual model import
# import your_floor_plan_model

def preprocess_parameters(
    room_count: int,
    square_footage: float,
    preferences: Dict[str, Any],
    additional_parameters: Optional[Dict[str, Any]] = None
) -> Dict[str, Any]:
    """
    Preprocess and validate the input parameters for the model
    """
    # Implement your preprocessing logic here
    processed_params = {
        "room_count": room_count,
        "square_footage": square_footage,
        "preferences": preferences,
    }
    
    if additional_parameters:
        processed_params.update(additional_parameters)
        
    return processed_params

def generate_floor_plan(
    room_count: int,
    square_footage: float,
    preferences: Dict[str, Any],
    additional_parameters: Optional[Dict[str, Any]] = None
) -> str:
    """
    Generate a floor plan using the AI model
    Returns a base64 encoded image
    """
    # Preprocess the parameters
    params = preprocess_parameters(
        room_count, 
        square_footage, 
        preferences, 
        additional_parameters
    )
    
    # This is where you would call your actual model
    # floor_plan = your_floor_plan_model.generate(**params)
    
    # For demonstration, we'll create a dummy image
    # In your actual implementation, this would be the output from your model
    dummy_image = Image.new('RGB', (800, 600), color='white')
    
    # Convert the image to base64 for API response
    buffered = io.BytesIO()
    dummy_image.save(buffered, format="PNG")
    img_str = base64.b64encode(buffered.getvalue()).decode()
    
    return img_str
