backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0,<2.0.0
python-multipart>=0.0.5
pillow>=8.3.1
numpy>=1.21.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=0.19.0
email-validator>=1.1.3
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.1
alembic>=1.7.1
python-dotenv>=0.19.0
# Add any other dependencies your AI model requires